+++ pipelines/processors/enhancement_processor.py
@@ -1,4 +1,5 @@
 import logging
+from typing import Callable, Any
 from pipelines.processors.base_processor import BaseProcessor
 from domain.exceptions import FileReadError, FileSaveError, ProcessorError
 
@@ -6,6 +7,14 @@ logger = logging.getLogger(__name__)
 
 class EnhancementProcessor(BaseProcessor):
     name = "enhancement"
+
+    def __init__(self):
+        super().__init__()
+
+        self.transform_steps: list[Callable[[Any, dict], Any]] = [
+            self._apply_clahe
+        ]
+
 
     def process(self, input_path: str, context: dict) -> str:
         import cv2
@@ -27,17 +36,28 @@ class EnhancementProcessor(BaseProcessor):
 
 
     def transform(self, image, context: dict):
+        for transform_step in self.transform_steps:
+            try:
+                image = transform_step(image, context)
+            except ProcessorError as e:
+                logger.error(f"ProcessorError from {self.name} processor ({transform_step.__name__}): {e}")
+                raise ProcessorError(f"ProcessorError from {self.name} processor ({transform_step.__name__}): {e}") from e
+
+        return image
+
+
+    def _apply_clahe(self, image, context: dict):
         import cv2
 
         clip_limit = context.get("clahe_clip_limit", 2.0)
-        tile_grid_size = context.get("clahe_tile_grid_size", (4,4))
+        tile_grid_size = context.get("clahe_tile_grid_size", (4, 4))
 
         clahe = cv2.createCLAHE(clipLimit=clip_limit, tileGridSize=tile_grid_size)
 
         try:
             result = clahe.apply(image)
         except Exception as e:
-            logger.error(f"Exception from apply CLAHE in transform: {e}")
-            raise ProcessorError(f"Exception from apply CLAHE in transform: {e}") from e
+            logger.error(f"Error when applying CLAHE: {e}")
+            raise ProcessorError(f"Error when applying CLAHE: {e}") from e
 
         return result